+++ plugins/systemd/SystemdRestart-inl.h
@@ -15,7 +15,7 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-#include "oomd/plugins/SystemdRestart.h"
+#include "oomd/plugins/systemd/SystemdRestart.h"
 
 #include <string>
